refactor: 패키지 분리
- domain: 도메인 내 객체 패키지
- handler: 입출력 핸들러 패키지
- message: 출력 메세지 패키지
- util: 유틸리티 패키지
