+++ src/main/java/racingcar/message/OutputMessage.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.message;
 
 public enum OutputMessage {
     CAR_NAME_INPUT_MESSAGE("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
