+++ src/test/java/racingcar/domain/CarsTest.java
@@ -1,14 +1,15 @@
-package racingcar;
+package racingcar.domain;
 
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static racingcar.Cars.MAX_CAR_SIZE;
-import static racingcar.Cars.MIN_CAR_SIZE;
-import static racingcar.ExceptionMessage.DUPLICATE_CAR_NAME;
-import static racingcar.ExceptionMessage.INVALID_SIZE_CAR_NAME;
+import static racingcar.domain.Cars.MAX_CAR_SIZE;
+import static racingcar.domain.Cars.MIN_CAR_SIZE;
+import static racingcar.message.ExceptionMessage.DUPLICATE_CAR_NAME;
+import static racingcar.message.ExceptionMessage.INVALID_SIZE_CAR_NAME;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
+
 class CarsTest {
     @Test
     void 자동차가_10개_초과면_예외가_발생한다() {
